Add session data to env vars
Currently the session secret can only be set by config.json or docker
secrets. This creates a problem on Heroku hosted instances that can not
set a session secret.

Since we automatcially generate them on startup this results in an
logout of all uers on every config change in Heroku.