Upgrade winston / refactor logging parts
Our log library got a new major version which should be implemented.

That's exactly what this patch does. Implementing the new version of the
logging library.

Along with the upgrade the new implementation of the logging library
provides an option to adjust the loglevel. This should be soon extended
to more debugging and verbose logging all over CodiMD. So in a long run
this should help debugging and improving the development.

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>