Update Reveal.JS to version 3.8.0
There was a new release of reval.js which requires minor changes on the
webpack configuration besides the simple update in the package.json.

Reveal.js implemented it's loader code into their main js file instead
of using the external head.min.js. Since, until now, we included this
using webpack, the absence of it broke the build process for the new
version. The changes in this pack remove the include of this file and
allows the usage of version 3.8.0.

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>